+++ src/pedestal/swagger/doc.clj
@@ -43,7 +43,7 @@
 (s/defschema Paths
   {s/Str {s/Keyword spec/Operation}})
 
-(s/defn doc-routes :- Paths
+(s/defn gen-paths :- Paths
   "Generates swagger paths from an expanded route table.
   This function can also be used to generate documentation
   offline or for easy debugging (turning schema checks on)."
@@ -55,7 +55,7 @@
            {path {method (apply deep-merge docs)}})))
 
 
-(defn inject-docs
+(s/defn inject-docs :- spec/Swagger
   "Attaches swagger information as a meta key to each documented
   route. The context passed to each interceptor has a reference to the
   selected route, so information like request and response schemas and
@@ -64,6 +64,6 @@
   (let [swagger-object (deep-merge {:swagger "2.0"
                                     :info {:title "Swagger API"
                                            :version "0.0.1"}
-                                    :paths (doc-routes route-table)}
+                                    :paths (gen-paths route-table)}
                                    docs)]
     (inject-swagger-into-routes route-table swagger-object)))