+++ test/pedestal/swagger/core_test.clj
@@ -5,9 +5,13 @@
             [schema.core :as s]
             [schema.test :as validation]
             [pedestal.swagger.doc :as doc]
+            [ring.util.response :refer [response]]
+            [ring.swagger.swagger2 :as spec]
             [io.pedestal.http.route.definition :as definition]
             [io.pedestal.http.body-params :as pedestal-body-params]
-            [io.pedestal.http :as bootstrap]))
+            [io.pedestal.http :as bootstrap]
+            [scjsv.core :as v]
+            [clojure.java.io :as io]))
 
 (use-fixtures :each validation/validate-schemas)
 
@@ -27,20 +31,18 @@
   {:summary "Put resource with id"
    :parameters {:body {:name s/Keyword}}}
   [{:keys [body-params path-params headers]}]
-  {:status 200 :headers {}
-   :body {:params (merge body-params path-params
-                         (select-keys headers [:auth]))}})
+  (response {:params (merge body-params path-params
+                            (select-keys headers [:auth]))}))
 
 (defhandler delete-handler
   {:summary "Delete resource with id"
    :parameters {:query {:notify s/Bool}}}
   [{:keys [query-params path-params headers]}]
-  {:status 200 :headers {}
-   :body {:params (merge query-params path-params
-                         (select-keys headers [:auth]))}})
+  (response {:params (merge query-params path-params
+                         (select-keys headers [:auth]))}))
 
 (defn non-documented-handler
-  [req] {:satus 200})
+  [req] (response {}))
 
 (defhandler get-handler
   {:summary "Get all resources"
@@ -62,7 +64,13 @@
       bootstrap/service-fn
       ::bootstrap/service-fn))
 
-(definition/defroutes routes
+(defn swagger-validator [swagger-object]
+  (let [validator (v/validator (slurp (io/resource "ring/swagger/v2.0_schema.json")))
+        json (spec/swagger-json swagger-object {})]
+    (validator json)
+    (response json)))
+
+ (definition/defroutes routes
   [["t" :test
     ["/" ^:interceptors [(pedestal-body-params/body-params)
                          (keywordize-params :headers) (body-params)
@@ -73,38 +81,40 @@
       {:put put-handler
        :delete delete-handler
        :head non-documented-handler}]
-     ["/doc" {:get [(swagger-doc)]}]]]])
-
-(deftest generates-correct-documentation
-  (let [expected {"/"
-                  {:get {:description "Requires auth as header"
-                         :summary "Get all resources"
-                         :parameters {:query {:q s/Str}
-                                      :header {:auth s/Str}}
-                         :responses {200 {:schema {:status s/Str}}
-                                     400 {}
-                                     500 {}
-                                     :default {:schema {:result [s/Str]}
-                                               :headers {(req "Location") s/Str}}}}}
-                  "/x/:id"
-                  {:put {:description "Requires id on path"
-                         :summary "Put resource with id"
-                         :parameters {:path {:id s/Int}
-                                      :header {:auth s/Str}
-                                      :body {:name s/Keyword}}
-                         :responses {400 {} 500 {}}}
-                   :delete {:description "Requires id on path"
-                            :summary "Delete resource with id"
-                            :parameters {:path {:id s/Int}
-                                         :header {:auth s/Str}
-                                         :query {:notify s/Bool}}
-                            :responses {400 {} 500 {}}}}}]
-    (is (= expected (doc/doc-routes routes)))))
+     ["/doc" {:get [(swagger-doc swagger-validator)]}]]]])
 
 (def app (make-app {::bootstrap/router :prefix-tree ;; or :linear-search
                     ::bootstrap/routes (doc/inject-docs
                                         {:title "Test"
                                          :version "0.1"} routes)}))
+
+(deftest generates-correct-paths
+  (let [paths {"/" {:get
+                    {:description "Requires auth as header"
+                     :summary "Get all resources"
+                     :parameters {:query {:q s/Str}
+                                  :header {:auth s/Str}}
+                     :responses {200 {:schema {:status s/Str}}
+                                 400 {}
+                                 500 {}
+                                 :default {:schema {:result [s/Str]}
+                                           :headers {(req "Location") s/Str}}}}}
+               "/x/:id" {:put
+                       {:description "Requires id on path"
+                        :summary "Put resource with id"
+                        :parameters {:path {:id s/Int}
+                                     :header {:auth s/Str}
+                                     :body {:name s/Keyword}}
+                        :responses {400 {} 500 {}}}
+                       :delete
+                       {:description "Requires id on path"
+                        :summary "Delete resource with id"
+                        :parameters {:path {:id s/Int}
+                                     :header {:auth s/Str}
+                                     :query {:notify s/Bool}}
+                        :responses {400 {}
+                                    500 {}}}}}]
+    (is (= paths (doc/gen-paths routes)))))
 
 (deftest coerces-params
   (are [resp req] (= resp (read-string (:body req)))
@@ -136,19 +146,33 @@
   (are [status resp req] (and (= status (:status req))
                               (= resp (read-string (:body req))))
        200 {:status "ok"}
-       (response-for app :get "http://t/?q=ok" :headers {"Auth" "y"})
+       (response-for app
+                     :get "http://t/?q=ok"
+                     :headers {"Auth" "y"})
 
        400 {:error {:query-params {:q "missing-required-key"}}}
-       (response-for app :get "http://t/" :headers {"Auth" "y"})
+       (response-for app
+                     :get "http://t/"
+                     :headers {"Auth" "y"})
 
        201 {:result ["a" "b"]}
-       (response-for app :get "http://t/?q=created" :headers {"Auth" "y"})
+       (response-for app
+                     :get "http://t/?q=created"
+                     :headers {"Auth" "y"})
 
        500 {:error {:headers {"Location" "missing-required-key"}
                     :body {:result "(not (sequential? \"fail\"))"}}}
-       (response-for app :get "http://t/?q=fail" :headers {"Auth" "y"})))
+       (response-for app
+                     :get "http://t/?q=fail"
+                     :headers {"Auth" "y"})))
 
 (deftest checks-swagger-handler-like-any-other-route
   (are [resp req] (= resp (read-string (:body req)))
        {:error {:headers {:auth "missing-required-key"}}}
        (response-for app :get "http://t/doc")))
+
+(deftest generates-valid-json-schema
+  (is (= 200 (:status
+              (response-for app
+                            :get "http://t/doc"
+                            :headers {"Auth" "y"})))))