+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.4.0
 - Bump pedestal to 0.4.0
+- Map parameter passed to defroutes is deep merged to the generated swagger object.
+- As a consequence `doc-routes` is properly renamed to `gen-paths`.
 
 ## 0.3.0
 - Initial version